medical3d_mvp/requirements.txt
PySide6>=6.7
numpy>=1.26
scipy>=1.11
scikit-image>=0.22
SimpleITK>=2.3
pydicom>=2.4
nibabel>=5.2
Pillow>=10.3
pyvista>=0.44
pyvistaqt>=0.11
vtk>=9.3
trimesh>=4.3
